Different app-wrapper
Create an app-wrapper div around everything but the header. The left and
right margins of the header aren't the same as the fluid margins of the
singularity "split" grid, so wrap the main content of the page in a new
div that will carry the grid system. The margins of this div can be kept
consistent with those of the header (if it isn't split)